Ignore "internal" globals when parsing lua
When we execute lua scripts to extract the global variables they set, we
are not interested in the things that already hang around in the globals
table when the Lua VM is initialized.

Since these things also contain binary that we don't want to run through
string decoding (resulting in `UnicodeDecodeError`), we're better off
getting rid of them.